+++ aobaker.h
@@ -15,12 +15,12 @@ namespace aobaker {
 	};
 
 	void BakeAoToVertices(
-		const float* firstVertexPosition, float* firstAoTarget, int vertexCount, size_t vertexStride, size_t targetStride,
+		const float* firstVertexPosition, float* firstAoTarget, int vertexCount, unsigned int vertexStride, unsigned int targetStride,
 		const unsigned int* indices, int indexCount,
 		const config& conf);
 
 	void BakeAoToVertices(
-		const double* firstVertexPosition, float* firstAoTarget, int vertexCount, size_t vertexStride, size_t targetStride,
+		const double* firstVertexPosition, float* firstAoTarget, int vertexCount, unsigned int vertexStride, unsigned int targetStride,
 		const unsigned int* indices, int indexCount,
 		const config& conf);
 }
